Move CLI key reading function up
Moves the `read_private_key` function from the circuit CLI up to the
`action` module so it can be used by other CLI commands.
